+++ src/ConfigProvider.php
@@ -32,10 +32,12 @@ class ConfigProvider
         // @codingStandardsIgnoreStart
         return [
             'aliases' => [
-                DEFAULT_DELEGATE     => Handler\NotFoundHandler::class,
-                DISPATCH_MIDDLEWARE  => Router\Middleware\DispatchMiddleware::class,
-                NOT_FOUND_MIDDLEWARE => Handler\NotFoundHandler::class,
-                ROUTE_MIDDLEWARE     => Router\Middleware\PathBasedRoutingMiddleware::class,
+                DEFAULT_DELEGATE            => Handler\NotFoundHandler::class,
+                DISPATCH_MIDDLEWARE         => Router\Middleware\DispatchMiddleware::class,
+                IMPLICIT_HEAD_MIDDLEWARE    => Router\Middleware\ImplicitHeadMiddleware::class,
+                IMPLICIT_OPTIONS_MIDDLEWARE => Router\Middleware\ImplicitOptionsMiddleware::class,
+                NOT_FOUND_MIDDLEWARE        => Handler\NotFoundHandler::class,
+                ROUTE_MIDDLEWARE            => Router\Middleware\PathBasedRoutingMiddleware::class,
             ],
             'factories' => [
                 Application::class                             => Container\ApplicationFactory::class,